requirements.txt
requests
random-word
telethon
deep-translator
gTTS
torch
transformers
python-dotenv